Fix: don't add duplicate value in possible values
In csp.Square#backtrack(), when adding values to the list of possible
values, I mistakenly added that value if it already exists in the list.
That added duplicate entries in the list and denied legit values which
were not in the list.

Now the value is added if it does NOT exists in the list.