Restructure: code SDF heuristic better
The method csp.Square#sdfH() first looped over all the cells and added
the unassigned cells in a temporary list and then it looped over that
list to find out the cell with the smallest domain.

The same is now done in one loop. Instead of creating a temporary list
of unassigned cells, the task of checking for the cell with smallest
domain is done in the first loop (the one that looped over all the cells
to find unassigned cells).