Fix: merge Forward Check and AC-3
The implementation of forward checking prior this commit was different
than that of ac3. But since ac3 is an extension of forward checking (it
just looks ahead), it makes sense to merge both of them together and use
a boolean flag to check if the extension is to be used or not.